+++ library/src/main/java/com/twilio/video/LocalVideoTrackStats.java
@@ -37,7 +37,8 @@ public class LocalVideoTrackStats extends LocalTrackStats {
      */
     public final int frameRate;
 
-    LocalVideoTrackStats(String trackId,
+    LocalVideoTrackStats(String trackSid,
+                         String trackId,
                          int packetsLost,
                          String codec,
                          String ssrc,
@@ -49,7 +50,7 @@ public class LocalVideoTrackStats extends LocalTrackStats {
                          VideoDimensions dimensions,
                          int capturedFrameRate,
                          int frameRate) {
-        super(trackId, packetsLost, codec, ssrc, timestamp,
+        super(trackSid, trackId, packetsLost, codec, ssrc, timestamp,
                 bytesSent, packetsSent, roundTripTime);
         this.captureDimensions = captureDimensions;
         this.dimensions = dimensions;