+++ library/src/main/java/com/twilio/video/RemoteVideoTrackStats.java
@@ -27,7 +27,8 @@ public class RemoteVideoTrackStats extends RemoteTrackStats {
      */
     public final int frameRate;
 
-    RemoteVideoTrackStats(String trackId,
+    RemoteVideoTrackStats(String trackSid,
+                          String trackId,
                           int packetsLost,
                           String codec,
                           String ssrc,
@@ -36,7 +37,7 @@ public class RemoteVideoTrackStats extends RemoteTrackStats {
                           int packetsReceived,
                           VideoDimensions dimensions,
                           int frameRate) {
-        super(trackId, packetsLost, codec, ssrc,
+        super(trackSid, trackId, packetsLost, codec, ssrc,
                 timestamp, bytesReceived, packetsReceived);
         this.dimensions = dimensions;
         this.frameRate = frameRate;